// types/auth.ts
export type PlatformRole = 
  | 'platform_super_admin'
  | 'platform_admin'
  | 'platform_manager'
  | 'platform_developer'
  | 'platform_customer_support'
  | 'platform_account_manager'
  | 'platform_financial_manager'
  | 'platform_content_moderator'
  | 'platform_data_analyst'
  | 'platform_operations_manager'
  | 'platform_outreach_manager'
  | 'platform_agent';

export type CompanyRole = 
  | 'b2c_company_owner'
  | 'b2c_company_admin'
  | 'b2c_marketing_director'
  | 'b2c_campaign_manager'
  | 'b2c_campaign_executive'
  | 'b2c_social_media_manager'
  | 'b2c_content_creator'
  | 'b2c_brand_manager'
  | 'b2c_performance_analyst'
  | 'b2c_finance_manager'
  | 'b2c_account_coordinator'
  | 'b2c_viewer';

export type InfluencerRole = 
  | 'influencer'
  | 'influencer_manager';

export type DetailedRole = PlatformRole | CompanyRole | InfluencerRole;